requirements.txt
asgiref==3.2.10
Django==3.1
django-autocomplete-light==3.5.1
django-filter==2.3.0
django-jquery==3.1.0
django-phonenumber-field==5.0.0
django-rest-multiple-models==2.1.3
djangorestframework==3.11.1
numpy==1.19.1
pandas==1.1.1
phonenumberslite==8.12.8
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
